Moving startup code to app.py, removing secret key
For secret key, just don’t upload real one to github
