+++ app/config.py
@@ -2,7 +2,4 @@ import os
 basedir = os.path.abspath(os.path.dirname(__file__))
 
 CRF_ENABLED = True
-SECRET_KEY = 'jdfaiwuhfi9ah49h4398u0rjg0a34'
-
-# SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, 'app.db')
-# SQLALCHEMY_MIGRATE_REPO = os.path.join(basedir, 'db_repository')
+SECRET_KEY = 'EXAMPLE_SECRET_KEY_HERE'
